tests/submission/Infraware/Forms/contents/Forms/fieldset_type.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>FieldSet_type</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<fieldset id="input_field">
 		</fieldset>
	</form>
	<script>

		var field = document.getElementById("input_field");
		var type = field.type;
		
		test(function() {
			assert_equals(type, "fieldset", "type attribute is not correct.");
		});

	</script>

 </body>
</html>
